// projects/bussiness-owners/src/app/modules/company/models/editbranchdto.ts
export interface EditBranchDto {
    id: string;
    branchName: string;
    branchRegion: string | null;
    branchCity: string | null;
    branchAddress: string | null;
    mobileNumberCode: string | null;
    mobileNumber: string | null;
    branchEmail: string | null;
    countryCode: string | null;
}